requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
pdfplumber
pytesseract
pdf2image
Pillow
openai
groq